requirements.txt
Django==3.2
django-allauth==0.44.0
django-filter==2.4.0
django-rest-auth==0.9.5
django-rest-authtoken==2.1.3
djangorestframework==3.12.4
asgiref==3.3.1
pytz==2021.1
sqlparse==0.4.1


